Address clang-tidy issues.
In cmake/AddCA.cmake, in add_ca_tidy, we try to add clang-tidy targets
for all source files, but for source files specified with relative
paths, we would resolve them using the wrong base directory. This change
uses file(REAL_PATH) with an explicit BASE_DIRECTORY to convert to an
absolute path and add the intended tidy rule.

Many files require updates to address the warnings we now see from
clang-tidy.

+++ source/cl/test/UnitCL/source/clCreateSubBuffer.cpp
@@ -26,7 +26,7 @@ class clCreateSubBufferTest : public ucl::CommandQueueTest {
                               sizeof(cl_uint), &mem_base_addr_align, nullptr));
     // NOTE: Ensure the buffer is large enough to create a sub buffer of any
     // scalar or vector type.
-    size = 3 * mem_base_addr_align;
+    size = 3 * static_cast<size_t>(mem_base_addr_align);
     region.origin = static_cast<size_t>(mem_base_addr_align);
     region.size = sizeof(cl_int);
   }
